# Allow Visibility to be inherited