# Objective

Fixes #4907. Fixes #838.
Supercedes #2087. Supercedes #865.

`Visibility` is currently entirely local. Set a parent entity to be invisible, and the children are still visible. This makes it hard for users to hide entire hierarchies of entities.

## Solution
Expand `Visibility` to hold both a "self" visibility, and a "inherited" visibility. The entity is visible if and only if both are true. A system similar to `transform_propagate_system` is then implemented to propagate visibility down through the hierarchy. Unlike Transforms, the action of setting visible or not is relatively light computationally so zero change detection is used for the system.

The `visibility_propagate_system` can run in parallel with `transform_propagate_system`, and is set as a dependency of `check_visibiility`, which will copy the same state to `ComputedVisibilty` once propagated. This structure is to avoid making a serialized chain of `transform_propagate` -> `check_visibiliity ` -> `visibility_propagate`. Due to the lighter nature of the visibility propagation, `visibility_propagate_system` will almost always finish faster than `transform_propagate_system` for perfectly overlapped hierarchies, so there is zero overhead on platforms that support multithreaded execution (basically everywhere except wasm32 platforms right now).

---

## Changelog
TODO

## Migration Guide
TODO