@LoopyAshy 2022-06-30
I was thinking about this very change recently when working on my project and I naively presumed adding `Visibility` & `ComputedVisibility` to the parent and setting it to false would make all the children invisible but this did not happen.