@james7132 2022-06-30
> Since `Visibility::inherited` is computed by the propagation system, shouldn't it be in `ComputedVisibility` ? Otherwise `Visibility` has a one field for the user to change and the other that shouldn't be touched.

It's either this or require ComputedVisibility for propagation, which isn't true for for some pipelines.