@mockersf 2022-06-30
The propagation system will always change the components, so it becomes impossible from the user to rely on `Changed<Visibility>`