@james7132 2022-06-30
> The propagation system will always change the components, so it becomes impossible from the user to rely on `Changed<Visibility>`

This is indeed a downside, though I guess we could put change detection back in to remedy this. My question is if we actually need change detection here. Even with this change, `Visibiilty` is still 4x smaller than `ComponentTicks`. It might actually be faster to iterate over them without change detection than rely on the filter.