@cart 2022-07-01
Change detection isn't just about optimization. It can be used for implementing logic. Ex: "send an event every time visibility changes for an entity"