@james7132 2022-07-01
> Change detection isn't just about optimization. It can be used for implementing logic. Ex: "send an event every time visibility changes for an entity"

OK, added a few small changes to then avoid triggering it when nothing has actually changed. This still keeps the same structure as before, but `Changed<Visibilty>` will trigger when either the local visibility or an ancestor's visibility has changed. Should make it easier to detect when an entire tree changes visibility.