@james7132 2022-07-01
@alice-i-cecile removing the Controversial tag since I think the only point that was of major concern was the change detection issue. Unless the overall design is of a concern, which I think might be quelled by the fact that cart suggested basically the same API [about a year ago](https://github.com/bevyengine/bevy/pull/2087#issuecomment-856370305).