@james7132 2022-07-02
> This looks good to me and I imagine you've done benchmarking to see that it doesn't significantly impact performance. There are some 'stress tests' for transform hierarchy that could probably be leveraged for testing if not.

Whoops, forgot to put those results up in the original one as perf justification for this approach.

Just to verify, I looked to ensure that the system parallelized with transform propagation, and blocks `check_visibility`, and it does. A slice from `many_cubes sphere`:

![image](https://user-images.githubusercontent.com/3137680/177015094-f294c163-fc7c-4245-8721-f39bcb759dff.png)

`many_cubes sphere` has a flat hierarchy with both transform and visibility on every entity. On it we see `transform_propagate` take 506.6us (mean), and `visibility_propagate` take 288.86us (mean).

For a stress test with deeper hierarchies, I tried `many_foxes` which has a deeper hierarchy. `transform_propagate` takes 1.44ms (mean) and `visibility_propagate` takes 770ns (mean).

Unfortunately I don't think we have many stress tests that target both in deep hierarchies.