@superdump 2022-07-04
There's the `transform_hierarchy` stress test that has a variety of configurations.