@james7132 2022-07-04
> There's the `transform_hierarchy` stress test that has a variety of configurations. And I was thinking more of a before/after comparison, though the breakdown of how much time `visibility_propagate` takes is at least promising for that part.

The visibility propagate currently requires visibility to be at every level of the hierarchy (like transforms): the transform hierarchy stress test wouldn't be a good test. I could try reworking it to support hierarchies with gaps, but this would be a huge hit to perf as it would likely match *all* entities, not just those bound to the hierarchy.