@cart 2022-07-12
I'm currently putting together the changes I outlined above (store computed hierarchy visibility on ComputedVisibility) so we can compare performance and make a call. This will also move sprites and ui to ComputedVisibility and will enable RenderLayers functionality. So it would also be an alternative to #5114 and #4007. I suspect it to perform worse than #5114 because it will be populating camera "visible entity" lists. But imo unifying these systems makes a lot of sense. Keeps the code smaller and the behaviors in sync / consistent. And if the current impl isn't good / fast enough for sprites, is it really good enough for 3d meshes? Unifying them would create pressure to optimize the system as a whole, which I like.