@cart 2022-07-12
Still working out implementation details, but heres some bevymark performance numbers (100,000 sprites) with various approaches:

* main: `0.017612s`
* PR: `0.017564s`
* Move hierarchy propagation to ComputedVisibility and port sprites, default them to "visible in view" and filter them out of `visibility_system` (skip view assignment, render layers, frustum cull option branching, etc): `0.017614s`
* Move hierarchy propagation to ComputedVisibility and port sprites (no other changes ... sprites are processed in `visibility_system`): `0.018317s`
  * note that this doesn't include porting sprite rendering to use the computed visibility lists. ill do this next

What this tells me is that the ComputedVisibility switch doesn't inherently change the cost (which makes a lot of sense to me), but adding render layer filter checks, checking for (non existent) frustum culling, and populating each view's "visible entity list" does have a cost. And that feels like a cost we should pay (but also try to minimize as much as possible).